requirements.txt
streamlit==1.24.0
nltk==3.8.1
transformers==4.30.2
torch==2.0.1
tokenizers==0.13.3
accelerate==0.20.3
sentencepiece==0.1.99
protobuf==3.20.0
textstat==0.7.3
pyperclip==1.8.2
python-docx==0.8.11 